Move data set preparation to separate class
